preprocess slides
